pyproject.toml
[project]
name = "nhp-aci"
dynamic = ["version"]
description = "New Hospital Programme Azure Container Instance helpers"

requires-python = ">=3.11"

dependencies = [
    "azure-identity>=1.12",
    "azure-mgmt-containerinstance>=10.1",
    "azure-storage-blob>=12.15",
    "jsonschema>=4.25",
    "python-dotenv>=1.0",
]


# Explicitly specify which packages to include
[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nhp.model.data.reference" = ["*.csv", "*.json"]

[project.optional-dependencies]
dev = [
    "coverage>=7.2",
    "pytest>=7.2",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "ruff>=0.11.10",
    "setuptools-scm>=8.3.1",
]

[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py311"
indent-width = 4
exclude = ["**/_version.py"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = [
"D", # pydocstyle
"PLC0415", # `import` should be at the top-level of a file
"PLR2004", # Magic value used in comparison
"PD901", # Avoid using the generic variable name `df`
"RUF005", # list concatenation
]

[tool.ruff.lint]
# Simple rules: pylint + isort
select = [
    "D",   # pydocstyle
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "I",   # isort (import sorting)
    "PL",  # pylint rules,
    "PD",  # pandas-vet
    "NPY", # NumPy-specific rules
    "RUF", # Ruff-specific rules
]

[tool.ruff.lint.isort]
# isort configuration
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 7

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"


[tool.pyright]
include = ["src", "tests"]
exclude = ["**/node_modules", "**/__pycache__"]
pythonVersion = "3.11"
typeCheckingMode = "basic"

[tool.setuptools_scm]
version_file = "src/nhp/_version.py"
